refactor *BeanMap to utils, adapt tests
