--- drools-core/src/test/java/org/drools/core/process/instance/impl/Person.java
@@ -1,31 +1,0 @@
-package org.drools.core.process.instance.impl;
-
-public class Person {
-
-    private String name;
-    private int age;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public int getAge() {
-        return age;
-    }
-
-    public void setAge(int age) {
-        this.age = age;
-    }
-
-    @Override
-    public String toString() {
-        return "Person{" +
-                "name='" + name + '\'' +
-                ", age=" + age +
-                '}';
-    }
-}
